Move networking related pkg/util to component-helpers
