+++ pkg/kubelet/dockershim/network/hostport/hostport_manager.go
@@ -30,10 +30,10 @@ import (
 
 	v1 "k8s.io/api/core/v1"
 	utilerrors "k8s.io/apimachinery/pkg/util/errors"
+	"k8s.io/component-helpers/networking/utils/conntrack"
+	utiliptables "k8s.io/component-helpers/networking/utils/iptables"
 	"k8s.io/klog/v2"
 	iptablesproxy "k8s.io/kubernetes/pkg/proxy/iptables"
-	"k8s.io/kubernetes/pkg/util/conntrack"
-	utiliptables "k8s.io/kubernetes/pkg/util/iptables"
 	"k8s.io/utils/exec"
 	utilnet "k8s.io/utils/net"
 )
