+++ pkg/kubelet/dockershim/network/hostport/hostport_manager_test.go
@@ -26,7 +26,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	v1 "k8s.io/api/core/v1"
-	utiliptables "k8s.io/kubernetes/pkg/util/iptables"
+	utiliptables "k8s.io/component-helpers/networking/utils/iptables"
 	"k8s.io/utils/exec"
 )
