+++ pkg/kubelet/dockershim/network/kubenet/kubenet_linux_test.go
@@ -30,13 +30,13 @@ import (
 	"github.com/stretchr/testify/mock"
 
 	utilsets "k8s.io/apimachinery/pkg/util/sets"
+	ipttest "k8s.io/component-helpers/networking/utils/iptables/testing"
 	kubeletconfig "k8s.io/kubernetes/pkg/kubelet/apis/config"
 	kubecontainer "k8s.io/kubernetes/pkg/kubelet/container"
 	"k8s.io/kubernetes/pkg/kubelet/dockershim/network"
 	mockcni "k8s.io/kubernetes/pkg/kubelet/dockershim/network/cni/testing"
 	nettest "k8s.io/kubernetes/pkg/kubelet/dockershim/network/testing"
 	"k8s.io/kubernetes/pkg/util/bandwidth"
-	ipttest "k8s.io/kubernetes/pkg/util/iptables/testing"
 	sysctltest "k8s.io/kubernetes/pkg/util/sysctl/testing"
 	"k8s.io/utils/exec"
 	fakeexec "k8s.io/utils/exec/testing"
