+++ pkg/kubelet/kubelet_network.go
@@ -19,10 +19,10 @@ package kubelet
 import (
 	"fmt"
 
-	"k8s.io/api/core/v1"
+	v1 "k8s.io/api/core/v1"
+	utiliptables "k8s.io/component-helpers/networking/utils/iptables"
 	runtimeapi "k8s.io/cri-api/pkg/apis/runtime/v1alpha2"
 	"k8s.io/klog/v2"
-	utiliptables "k8s.io/kubernetes/pkg/util/iptables"
 )
 
 const (
