+++ pkg/proxy/ipvs/proxier_test.go
@@ -35,6 +35,12 @@ import (
 	"k8s.io/apimachinery/pkg/util/sets"
 	utilfeature "k8s.io/apiserver/pkg/util/feature"
 	featuregatetesting "k8s.io/component-base/featuregate/testing"
+	utilipset "k8s.io/component-helpers/networking/utils/ipset"
+	ipsettest "k8s.io/component-helpers/networking/utils/ipset/testing"
+	utiliptables "k8s.io/component-helpers/networking/utils/iptables"
+	iptablestest "k8s.io/component-helpers/networking/utils/iptables/testing"
+	utilipvs "k8s.io/component-helpers/networking/utils/ipvs"
+	ipvstest "k8s.io/component-helpers/networking/utils/ipvs/testing"
 	"k8s.io/kubernetes/pkg/features"
 	"k8s.io/kubernetes/pkg/proxy"
 	"k8s.io/kubernetes/pkg/proxy/healthcheck"
@@ -42,12 +48,6 @@ import (
 	proxyutiliptables "k8s.io/kubernetes/pkg/proxy/util/iptables"
 	proxyutiltest "k8s.io/kubernetes/pkg/proxy/util/testing"
 	"k8s.io/kubernetes/pkg/util/async"
-	utilipset "k8s.io/kubernetes/pkg/util/ipset"
-	ipsettest "k8s.io/kubernetes/pkg/util/ipset/testing"
-	utiliptables "k8s.io/kubernetes/pkg/util/iptables"
-	iptablestest "k8s.io/kubernetes/pkg/util/iptables/testing"
-	utilipvs "k8s.io/kubernetes/pkg/util/ipvs"
-	ipvstest "k8s.io/kubernetes/pkg/util/ipvs/testing"
 	"k8s.io/utils/exec"
 	fakeexec "k8s.io/utils/exec/testing"
 	utilpointer "k8s.io/utils/pointer"
