+++ pkg/proxy/ipvs/safe_ipset.go
@@ -19,7 +19,7 @@ package ipvs
 import (
 	"sync"
 
-	"k8s.io/kubernetes/pkg/util/ipset"
+	"k8s.io/component-helpers/networking/utils/ipset"
 )
 
 type safeIpset struct {
