+++ staging/src/k8s.io/component-helpers/networking/utils/ipset/testing/fake.go
@@ -20,7 +20,7 @@ import (
 	"fmt"
 
 	"k8s.io/apimachinery/pkg/util/sets"
-	"k8s.io/kubernetes/pkg/util/ipset"
+	"k8s.io/component-helpers/networking/utils/ipset"
 )
 
 // FakeIPSet is a no-op implementation of ipset Interface
