+++ staging/src/k8s.io/component-helpers/networking/utils/ipvs/testing/fake.go
@@ -22,7 +22,7 @@ import (
 	"strconv"
 	"time"
 
-	utilipvs "k8s.io/kubernetes/pkg/util/ipvs"
+	utilipvs "k8s.io/component-helpers/networking/utils/ipvs"
 )
 
 //FakeIPVS no-op implementation of ipvs Interface
